% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcoholsmoke}
\alias{alcoholsmoke}
\title{Alcohol Consumption and Smoking Status}
\format{
A data frame with 863 rows and 2 variables:
\describe{
\item{alcohol_drinks}{Number of alcoholic drinks consumed per week}
\item{smoked}{Whether the respondent has ever smoked:
\code{0} = No, \code{1} = Yes}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
alcoholsmoke
}
\description{
Self-reported weekly alcohol consumption and smoking status from a
large health survey of 863 respondents. Useful for exploring the
association between a quantitative variable (drinks per week) and a
binary coded variable (ever smoked). Chapter 5 and 10 of ISI.
}
\note{
The \code{smoked} variable is coded numerically (0/1). Convert
to a character variable for use with two-proportion functions:
\code{alcoholsmoke$smoked_f <- ifelse(alcoholsmoke$smoked == 1, "Yes", "No")}
}
\examples{
head(alcoholsmoke)
alcoholsmoke$smoked_f <- ifelse(alcoholsmoke$smoked == 1, "Yes", "No")
test_1prop(
  formula       = ~ smoked_f,
  data          = alcoholsmoke,
  success_level = "Yes",
  null_pi       = 0.25,
  alternative   = "two.sided",
  method        = "theory"
)
}
\keyword{datasets}
