% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auction}
\alias{auction}
\title{Dutch vs. First-Price Auction Bids}
\format{
A data frame with 88 rows and 2 variables:
\describe{
\item{dutch}{Bid amount in the Dutch auction}
\item{fp}{Bid amount in the first-price sealed-bid auction}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
auction
}
\description{
Bids placed by the same participants in two auction formats -- a Dutch
(descending price) auction and a first-price sealed-bid auction.
Because the same person bid in both formats, this is a paired design.
Used in Chapter 7 of ISI for paired inference.
}
\examples{
head(auction)
test_paired(
  formula     = fp ~ dutch,
  data        = auction,
  alternative = "two.sided",
  method      = "theory"
)
ci_paired(
  formula    = fp ~ dutch,
  data       = auction,
  conf_level = 0.95,
  method     = "theory"
)
}
\keyword{datasets}
