% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biketimes}
\alias{biketimes}
\title{Bike Frame Type and Ride Times}
\format{
A data frame with 56 rows and 2 variables:
\describe{
\item{frame}{Bike frame type: \code{"carbon"} or \code{"steel"}}
\item{time}{Ride time (minutes)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
biketimes
}
\description{
Ride times for cyclists using two different bike frame types --
carbon and steel. Used in Chapter 6 of ISI for two-mean hypothesis
tests and confidence intervals. This is a randomized experiment,
making causal conclusions appropriate.
}
\examples{
head(biketimes)
explore_2vars(formula = time ~ frame, data = biketimes)
test_2mean(
  formula     = time ~ frame,
  data        = biketimes,
  alternative = "two.sided",
  method      = "theory"
)
ci_2mean(
  formula    = time ~ frame,
  data       = biketimes,
  conf_level = 0.95,
  method     = "theory"
)
}
\keyword{datasets}
