% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood}
\alias{blood}
\title{Blood Donor Response by Year}
\format{
A data frame with 2698 rows and 2 variables:
\describe{
\item{year}{Year the survey was conducted: \code{2002} or \code{2004}}
\item{response}{Participant response:
\code{"donated"} or \code{"did.not"}}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
blood
}
\description{
Survey responses about blood donation collected in two years (2002
and 2004). Used in Chapter 5 of ISI to compare donation rates across
years using two-proportion inference.
}
\examples{
head(blood)
test_2prop(
  formula       = response ~ year,
  data          = blood,
  success_level = "donated",
  alternative   = "two.sided",
  method        = "theory"
)
}
\keyword{datasets}
