% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bowlsmms}
\alias{bowlsmms}
\title{M&Ms Eaten from Small vs. Large Bowl}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{small}{Number of M&Ms eaten from the small bowl}
\item{large}{Number of M&Ms eaten from the large bowl}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
bowlsmms
}
\description{
Number of M&Ms eaten by participants when served from a small versus
a large bowl. Each participant experienced both conditions, making
this a paired design. Used in Chapter 7 of ISI for paired inference.
A fun example illustrating how container size affects consumption.
}
\examples{
head(bowlsmms)
test_paired(
  formula     = large ~ small,
  data        = bowlsmms,
  alternative = "greater",
  method      = "theory"
)
}
\keyword{datasets}
