% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastfeedintell}
\alias{breastfeedintell}
\title{Breastfeeding and Child Intelligence}
\format{
A data frame with 322 rows and 2 variables:
\describe{
\item{feeding}{Feeding method:
\code{"Breastfed"} or \code{"NotBreastfed"}}
\item{gci}{General Cognitive Index score (higher = better)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
breastfeedintell
}
\description{
A study comparing General Cognitive Index (GCI) scores between
children who were breastfed and those who were not. Used in Chapter 6
of ISI for two-mean inference. A good example for discussing
confounding and the distinction between causation and association.
}
\examples{
head(breastfeedintell)
test_2mean(
  formula     = gci ~ feeding,
  data        = breastfeedintell,
  alternative = "two.sided",
  method      = "theory"
)
ci_2mean(
  formula    = gci ~ feeding,
  data       = breastfeedintell,
  conf_level = 0.95,
  method     = "theory"
)
}
\keyword{datasets}
