% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{closefriends}
\alias{closefriends}
\title{Number of Close Friends by Sex}
\format{
A data frame with 1467 rows and 2 variables:
\describe{
\item{sex}{\code{"Men"} or \code{"Women"}}
\item{friends}{Number of close friends reported}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
closefriends
}
\description{
Survey data on the number of close friends reported by male and female
respondents. Used in Chapter 6 of ISI for two-mean inference with a
large real-world sample. A useful example for discussing skewness and
validity conditions.
}
\examples{
head(closefriends)
test_2mean(
  formula     = friends ~ sex,
  data        = closefriends,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
