% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dolphin}
\alias{dolphin}
\title{Dolphin-Assisted Therapy}
\format{
A data frame with 30 rows and 2 variables:
\describe{
\item{swimming}{Treatment assignment:
\code{"Dolphin"} or \code{"Control"}}
\item{response}{Outcome:
\code{"Improve"} or \code{"NotImprove"}}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
dolphin
}
\description{
Results from a randomized experiment testing whether swimming with
dolphins improves outcomes for patients with mild-to-moderate
depression. Participants were assigned to either swim with dolphins
or swim without them (control). Used in Chapter 5 of ISI for
two-proportion inference.
}
\examples{
head(dolphin)
test_2prop(
  formula       = response ~ swimming,
  data          = dolphin,
  success_level = "Improve",
  alternative   = "greater",
  method        = "simulation"
)
ci_2prop(
  formula       = response ~ swimming,
  data          = dolphin,
  success_level = "Improve",
  conf_level    = 0.95,
  method        = "theory"
)
}
\keyword{datasets}
