% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{draftlottery}
\alias{draftlottery}
\title{Vietnam Draft Lottery}
\format{
A data frame with 366 rows and 2 variables:
\describe{
\item{sequential_date}{Day of the year (1 = January 1,
366 = December 31)}
\item{draft_number}{Assigned draft number (1 = called first,
366 = called last)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
draftlottery
}
\description{
Draft numbers assigned in the 1969 Vietnam War lottery, matched to
sequential birth date (day of year 1-366). If the lottery was truly
random, there should be no correlation between birth date and draft
number. Used in Chapter 10 of ISI to test that randomness claim via
correlation inference.
}
\examples{
head(draftlottery)
explore_2vars(
  formula  = draft_number ~ sequential_date,
  data     = draftlottery,
  fit_line = TRUE
)
test_correlation(
  formula     = draft_number ~ sequential_date,
  data        = draftlottery,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
