% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examtimesscores}
\alias{examtimesscores}
\title{Exam Time and Scores}
\format{
A data frame with 30 rows and 2 variables:
\describe{
\item{time}{Time spent on the exam (minutes)}
\item{score}{Exam score (points)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
examtimesscores
}
\description{
Time spent on an exam and the resulting score for 30 students. Used
in Chapter 10 of ISI to explore whether spending more time on an exam
is associated with a higher (or lower) score. Results often surprise
students and prompt good discussion about confounding.
}
\examples{
head(examtimesscores)
explore_2vars(
  formula  = score ~ time,
  data     = examtimesscores,
  fit_line = TRUE
)
test_regression(
  formula     = score ~ time,
  data        = examtimesscores,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
