% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exercisemood}
\alias{exercisemood}
\title{Exercise Intensity and Mood Change}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{exercise_intensity}{Exercise intensity level (numeric)}
\item{change_mood}{Change in mood score after exercise (positive =
mood improved, negative = mood worsened)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
exercisemood
}
\description{
Change in mood score after exercise at varying intensity levels for 32
participants. Used in Chapter 10 of ISI for regression and correlation.
Explores whether higher exercise intensity is associated with greater
mood improvement.
}
\examples{
head(exercisemood)
explore_2vars(
  formula  = change_mood ~ exercise_intensity,
  data     = exercisemood,
  fit_line = TRUE
)
test_regression(
  formula     = change_mood ~ exercise_intensity,
  data        = exercisemood,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
