% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_2vars.R
\name{explore_2vars}
\alias{explore_2vars}
\title{Explore Bivariate Data (Correlation and Simple Linear Regression)}
\usage{
explore_2vars(x, y, group = NULL, x_name = "X", y_name = "Y", fit_line = FALSE)
}
\arguments{
\item{x}{Numeric vector for the explanatory variable}

\item{y}{Numeric vector for the response variable}

\item{group}{Optional categorical vector for grouping}

\item{x_name}{Character string for the X-axis label}

\item{y_name}{Character string for the Y-axis label}

\item{fit_line}{Logical; FALSE for Correlation mode, TRUE for Regression mode}
}
\value{
A ggplot2 or patchwork object
}
\description{
Explore Bivariate Data (Correlation and Simple Linear Regression)
}
