% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facebook}
\alias{facebook}
\title{Facebook Friends and Network Density}
\format{
A data frame with 40 rows and 2 variables:
\describe{
\item{friends}{Number of Facebook friends}
\item{density}{Network density score}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
facebook
}
\description{
Number of Facebook friends and network density scores for 40 users.
Network density measures how interconnected a user's friend network is.
Used in Chapter 10 of ISI for regression and correlation.
}
\examples{
head(facebook)
explore_2vars(formula = density ~ friends, data = facebook)
test_correlation(
  formula     = density ~ friends,
  data        = facebook,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
