% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{footheight}
\alias{footheight}
\title{Foot Length and Height}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{footlength}{Foot length (cm)}
\item{height}{Height (inches)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
footheight
}
\description{
Foot length and height measurements from a sample of 20 individuals.
A simple, intuitive dataset for introducing regression and correlation
in Chapter 10 of ISI. Students can easily reason about the expected
direction of the relationship before running the analysis.
}
\examples{
head(footheight)
explore_2vars(formula = height ~ footlength, data = footheight)
test_regression(
  formula     = height ~ footlength,
  data        = footheight,
  alternative = "greater",
  method      = "theory"
)
}
\keyword{datasets}
