% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goodandbad}
\alias{goodandbad}
\title{Question Wording and Perception}
\format{
A data frame with 29 rows and 2 variables:
\describe{
\item{wording}{Question framing: contains values like
\code{"\"goodyear\""} or \code{"\"badyear\""}}
\item{perception}{Participant response:
\code{"positive"} or \code{"negative"}}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
goodandbad
}
\description{
A study examining whether the wording of a survey question -- framing
it around a "good year" vs. a "bad year" -- affects participants'
perceptions of economic conditions. Used in Chapter 5 of ISI for
two-proportion inference.
}
\examples{
head(goodandbad)
table(goodandbad$wording, goodandbad$perception)
}
\keyword{datasets}
