% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haircuts}
\alias{haircuts}
\title{Haircut Costs by Sex}
\format{
A data frame with 50 rows and 2 variables:
\describe{
\item{sex}{\code{"Male"} or \code{"Female"}}
\item{cost}{Cost of the most recent haircut (dollars)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
haircuts
}
\description{
Haircut costs reported by male and female college students. Used in
both Chapter 2 (one-mean) and Chapter 6 (two-mean) of ISI. A
consistently engaging dataset that generates real class discussion
about whether men and women pay different amounts for haircuts.
}
\examples{
head(haircuts)
# One-mean: is the average haircut cost different from $50?
test_1mean(
  formula     = ~ cost,
  data        = haircuts,
  null_mu     = 50,
  alternative = "two.sided",
  method      = "theory"
)
# Two-mean: do men and women pay different amounts?
test_2mean(
  formula     = cost ~ sex,
  data        = haircuts,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
