% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{handwidth}
\alias{handwidth}
\title{Hand Width and Perceived Weight}
\format{
A data frame with 46 rows and 2 variables:
\describe{
\item{hand_width}{Hand width (cm)}
\item{perceived_weight}{Perceived weight rating (higher = heavier)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
handwidth
}
\description{
Hand width measurements and perceived weight ratings from a study on
sensory perception. Used in Chapter 10 of ISI for regression and
correlation. Explores whether people with wider hands perceive objects
as heavier.
}
\examples{
head(handwidth)
explore_2vars(formula = perceived_weight ~ hand_width, data = handwidth)
test_correlation(
  formula     = perceived_weight ~ hand_width,
  data        = handwidth,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
