% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jjvsbicycle}
\alias{jjvsbicycle}
\title{Jumping Jacks vs. Bicycle Calories Burned}
\format{
A data frame with 22 rows and 2 variables:
\describe{
\item{jj}{Calories burned doing jumping jacks}
\item{bicycle}{Calories burned riding a stationary bicycle}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
jjvsbicycle
}
\description{
Calories burned by participants doing jumping jacks versus riding a
stationary bicycle. Each participant completed both exercises, making
this a paired design. Used in Chapter 7 of ISI to introduce the
paired T-test.
}
\examples{
head(jjvsbicycle)
test_paired(
  formula     = bicycle ~ jj,
  data        = jjvsbicycle,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
