% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oldfaithful2}
\alias{oldfaithful2}
\title{Old Faithful Eruption Type and Wait Times}
\format{
A data frame with 222 rows and 2 variables:
\describe{
\item{eruption_type}{Type of preceding eruption:
\code{"short"} or \code{"long"}}
\item{time_between}{Wait time until the next eruption (minutes)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
oldfaithful2
}
\description{
Wait times between eruptions of the Old Faithful geyser, categorized
by eruption type. Used in Chapter 6 of ISI to compare wait times
between short and long eruptions using two-mean inference. This is
the richer version of the dataset -- it includes the grouping variable
that \code{oldfaithful} lacks.
}
\examples{
head(oldfaithful2)
explore_2vars(formula = time_between ~ eruption_type, data = oldfaithful2)
test_2mean(
  formula     = time_between ~ eruption_type,
  data        = oldfaithful2,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
