% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{platesize}
\alias{platesize}
\title{Dinner Plate Size Over Time}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{year}{Year of measurement}
\item{size}{Average plate diameter (inches)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
platesize
}
\description{
Average dinner plate diameter recorded across different years, from
1950 onward. Used in Chapter 10 of ISI to discuss trends over time
and simple linear regression. Plates have gotten larger over the
decades -- but is the trend statistically significant?
}
\examples{
head(platesize)
explore_2vars(formula = size ~ year, data = platesize, fit_line = TRUE)
test_regression(
  formula     = size ~ year,
  data        = platesize,
  alternative = "greater",
  method      = "theory"
)
table_regression(formula = size ~ year, data = platesize)
}
\keyword{datasets}
