% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleep}
\alias{sleep}
\title{Sleep Deprivation and Reaction Time}
\format{
A data frame with 21 rows and 2 variables:
\describe{
\item{sleep}{Sleep condition:
\code{"deprived"} or \code{"unrestricted"}}
\item{time}{Change in reaction time score (higher = better)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
sleep
}
\description{
Change in reaction time scores for participants assigned to either
a sleep-deprived condition or an unrestricted sleep condition. Used in
Chapter 6 of ISI as a core example for two-mean inference. Positive
values indicate improvement; negative values indicate worsening.
}
\examples{
head(sleep)
test_2mean(
  formula     = time ~ sleep,
  data        = sleep,
  alternative = "less",
  method      = "theory"
)
}
\keyword{datasets}
