% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempheart}
\alias{tempheart}
\title{Body Temperature and Heart Rate}
\format{
A data frame with 130 rows and 2 variables:
\describe{
\item{body_temp}{Body temperature (degrees Fahrenheit)}
\item{heart_rate}{Resting heart rate (beats per minute)}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
tempheart
}
\description{
Body temperature and resting heart rate measurements from 130
individuals. A classic dataset used in Chapter 10 of ISI for regression
and correlation. Also widely used for one-mean inference on body
temperature (testing whether the true mean is 98.6 degrees F).
}
\examples{
head(tempheart)

# One-mean: is the average body temperature really 98.6 F?
test_1mean(
  formula     = ~ body_temp,
  data        = tempheart,
  null_mu     = 98.6,
  alternative = "two.sided",
  method      = "theory"
)

# Correlation
explore_2vars(formula = heart_rate ~ body_temp, data = tempheart)
test_correlation(
  formula     = heart_rate ~ body_temp,
  data        = tempheart,
  alternative = "two.sided",
  method      = "theory"
)
}
\keyword{datasets}
