% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeestimate}
\alias{timeestimate}
\title{Time Interval Estimates (Sample)}
\format{
A data frame with 48 rows and 1 variable:
\describe{
\item{estimate}{Estimated time interval in seconds}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
timeestimate
}
\description{
Students were asked to estimate a 10-second time interval without
counting. Used in Chapter 2 of ISI to introduce one-mean inference --
specifically, to test whether students systematically over- or
underestimate a 10-second interval.
}
\examples{
head(timeestimate)
test_1mean(
  formula     = ~ estimate,
  data        = timeestimate,
  null_mu     = 10,
  alternative = "two.sided",
  method      = "theory"
)
ci_1mean(
  formula    = ~ estimate,
  data       = timeestimate,
  conf_level = 0.95,
  method     = "theory"
)
}
\keyword{datasets}
