% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timepopulation}
\alias{timepopulation}
\title{Time Interval Estimates (Population)}
\format{
A data frame with 6215 rows and 1 variable:
\describe{
\item{estimate}{Estimated time interval in seconds}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
timepopulation
}
\description{
The full population of time interval estimates from a large-scale
version of the time estimation activity. Paired with \code{timeestimate},
this dataset is used in Chapter 2 of ISI to illustrate the distinction
between a sample and its population -- a foundational concept for
understanding sampling variability.
}
\examples{
head(timepopulation)
# Compare sample mean vs population mean
mean(timeestimate$estimate)
mean(timepopulation$estimate)
}
\keyword{datasets}
