% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yawning}
\alias{yawning}
\title{Yawning Contagion Experiment}
\format{
A data frame with 50 rows and 2 variables:
\describe{
\item{yawn_seed}{Whether the participant saw someone yawn:
\code{"Seeded"} or \code{"Control"}}
\item{response}{Whether the participant yawned:
\code{"Yawn"} or \code{"NoYawn"}}
}
}
\source{
\url{http://www.isi-stats.com/isi/}
}
\usage{
yawning
}
\description{
Data from a MythBusters episode testing whether yawning is contagious.
Participants were either exposed to a yawn seed or placed in a control
condition, then observed for whether they yawned. Used in Chapter 5 of
ISI for two-proportion inference and to discuss randomization.
}
\examples{
head(yawning)
test_2prop(
  formula       = response ~ yawn_seed,
  data          = yawning,
  success_level = "Yawn",
  alternative   = "two.sided",
  method        = "simulation"
)
}
\keyword{datasets}
