% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_regression.R
\name{test_regression}
\alias{test_regression}
\title{Simple Linear Regression Hypothesis Test (Slope)}
\usage{
test_regression(
  x,
  y,
  alternative = "two.sided",
  method = "theory",
  sim_reps = 1000
)
}
\arguments{
\item{x}{Numeric vector for the explanatory variable}

\item{y}{Numeric vector for the response variable}

\item{alternative}{"two.sided", "greater", or "less"}

\item{method}{"theory" or "simulation"}

\item{sim_reps}{Number of iterations for Simulation (default 1000)}
}
\value{
An S3 object of class stat218_test_regression
}
\description{
Simple Linear Regression Hypothesis Test (Slope)
}
